+++ common.py
@@ -1,0 +1,1 @@
+asset_folder = './assets/'
